fix(autoshutdown): revert to nodeps install
As requested by the team, reverting from full install + UI disable,
to nodeps install. This maybe makes things more brittle in case any
dependencies are introduced in the extension that aren't already
available in the SMStudio image, but does seem to make the script
faster to run which is nice.